Add files via upload
Let the user be asked for cinema or theater preference. 15 TL to watch cinema, 10 TL to watch theater
does not have to be paid. Considering that there is a 50% discount for students, there is a discount if you are a student.
made; Python code that calculates the non-discounted amount and prints it to the screen if the student is not a student