Add files via upload
A histogram is a type of chart used to show the distribution of a continuous variable. variable
value ranges (bins) are determined and how many samples are in each range are visually represented.
The histogram is used to see the general shape, central tendency, and spread of the data distribution.
We use pandas' plot.hist() function to create a histogram chart.