Add files via upload
Dot Plot or Scatter Plot to visualize the relationship or distribution between two continuous variables
It is a type of chart used. Each data point represents the value of one variable on one axis and the other
On the axis it represents the value of the other variable. Scatter plots show the relationship between variables.
It is used to analyze data, see correlation, or detect outliers.
We use pandas' plot.scatter() function to create a dot plot.