Add files via upload
Dot Plot or Scatter Plot to visualize the relationship or distribution between two continuous variables
It is a type of chart used. Each data point represents the value of one variable on one axis and the other
On the axis it represents the value of the other variable. Scatter plots show the relationship between variables.
It is used to analyze data, see correlation, or detect outliers.
Box Plot or Box-Whisker Plot (Box Plot) shows the central tendency, spread and outliers of the data.
It is a type of chart used to visualize values. Boxplot shows data at minimum, first
a box representing the quartile (Q1), median (Q2), third quartile (Q3), and maximum values, and
It is indicated by the "whiskers" on the top and bottom of the box.
To create a box plot, we use pandas' plot.box() function.