Add files via upload
Pie Chart, visualizing the proportions of different parts of a whole
It is a type of chart used for . A pie chart is used to show the percentage distribution of data. Each
A data point is represented as a slice, and the sum of all slices equals 100%.
We use pandas' plot.pie() function to create a pie chart.