Add files via upload
Area Plot or Area Plot is a plot of a variable plotted over time or another independent variable.
It is a type of chart used to visualize how things change depending on the situation. Area chart of variable
It shows how values ​​spread across time or another axis and overlap each other. This
chart type for displaying components within a whole, such as the sum or percentage contributions of a variable
can also be used.
We use pandas' plot.area() function to create an area plot.