Add files via upload
Density Plot or Density Plot shows the general shape of the data distribution and
It is a type of chart used to visualize density estimation. Density plot of the data set
It is plotted as a curve reflecting the probability density. This graph is used to better understand the distribution of the data set.
and is used to identify possible peaks or dense regions.
We use the seaborn library to create density plots.