Add files via upload
Time series analysis is a statistical method that involves the analysis and prediction of time-varying data.
Time series are data measured at regular intervals and are measured in future values.
It has a structure that can be used to estimate values. Pandas library, time series
It provides a set of special functions and data structures for analysis.