Add files via upload
• Ask the user to enter a number between 0-20 (20 included).
• If the user makes an inappropriate entry, the specified
Give a warning message so that the user can log in according to the conditions.
• If the number entered by the user is odd, the factorial of the entered number is calculated and displayed on the screen.
print:
For example, if the entered number is 5, the output to be seen on the screen is: 5! = 120
• If the number entered by the user is even, it starts from 0 to the number entered by the user (the entered number
The sum of even numbers (including) is written on the screen.
For example, if the entered number is 12, the output to be seen on the screen is:
The sum is 42. Write the Python code.