Add files via upload
Level number guessing game;
 • 2 playing modes are available to the user, called easy and hard.
will be offered.
• The user who chooses the hard mode has 5 rights, and the user who chooses the easy mode has 10 rights.
will be.
• If the user guesses correctly, “You Won the Game!” appears on the screen. print.
• If the user's estimate is greater than the computer's estimate, "Choose a smaller number"
guess!” print the text on the screen and the user guesses the next one, if any.
let him do it
• If the user's estimate is smaller than the computer's estimate, “A larger
Guess the number!” print the text on the screen and if there is a user, the next
Let him make his guess. Write the Python code.